requirements.txt
pillow
pyautogui
pydirectinput
pynput
